refactoring import to move the rawIdentity stuff to its own class.
